Move schema definition to separate module and update Astro config
