+++ astro.config.mjs
@@ -2,34 +2,10 @@ import { defineConfig } from "astro/config";
 import tailwind from "@astrojs/tailwind";
 import netlify from "@astrojs/netlify";
 import react from "@astrojs/react";
-import { em, entity, text, number, date } from "bknd/data";
 import { randomString } from "bknd/utils";
 
+import { schema } from "./src/lib/bknd";
 import { addBknd } from "./src/integrations/bknd";
-
-const schema = em(
-  {
-    posts: entity("posts", {
-      // "id" is automatically added
-      title: text().required(),
-      slug: text().required(),
-      publishDate: date().required(),
-      content: text(),
-      views: number()
-    }),
-    comments: entity("comments", {
-      content: text()
-    })
-
-    // relations and indices are defined separately.
-    // the first argument are the helper functions, the second the entities.
-  },
-  ({ relation, index }, { posts, comments }) => {
-    relation(comments).manyToOne(posts);
-    // relation as well as index can be chained!
-    index(posts).on(["title"]).on(["slug"], true);
-  }
-);
 
 const bkndConfig = {
   adminRoute: "/admin",
